requirements.txt
django-heroku==0.3.1
numpy==1.20.2
Django==3.2.4
django-phonenumber-field==5.2.0
django-cors-headers==3.7.0
django-allauth==0.44.0
django-address==0.2.5
django-rest-auth==0.9.5
djangorestframework==3.12.4
dj-database-url==0.5.0
python-dotenv==0.17.1
phonenumbers==8.12.21
psycopg2==2.8.6
gunicorn==20.1.0
whitenoise==5.2.0
stripe==2.58.0